Moved current redux items into appropriate directory
