--- frontend/src/actions.js
@@ -1,2 +1,0 @@
-const REVIEW_ADDED = 'reviewAdded';
-const REVIEW_DELETED = 'reviewDeleted';
