+++ frontend/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import store from './store';
+import store from './redux/store';
 import { Provider } from 'react-redux';
