+++ frontend/src/redux/actions.js
@@ -1,0 +1,4 @@
+const REVIEW_ADDED = 'reviewAdded';
+const REVIEW_DELETED = 'reviewDeleted';
+
+const CURRENT_USER_STORED = 'loggedIn';
