+++ frontend/src/redux/store.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux';
-import reviewReducer from './reviewReducer';
+import reviewReducer from './reducers/reviewReducer';
 
 
 const store = createStore(reviewReducer);
